requirements.txt
Flask
tensorflow-macos
opencv-python
numpy
Pillow
matplotlib